+++ config/webpack.prod.js
@@ -7,16 +7,13 @@ const buildVersion = function() {
   return require("../package.json").version;
 };
 
-const babel = require("./babel");
-babel.presets.push("babel-preset-minify");
-
 module.exports = merge(common, {
+  mode: "production",
   devtool: "source-map",
   plugins: [
     new webpack.DefinePlugin({
       "process.env.NODE_ENV": JSON.stringify("production"),
-      "process.env.VERSION": JSON.stringify(buildVersion()),
-      "process.env.GA_TRACKER": JSON.stringify("UA-71899181-1")
+      "process.env.VERSION": JSON.stringify(buildVersion())
     }),
     new UglifyJSPlugin(uglifyConfig)
   ]