+++ src/editor/components/ScrollBox.js
@@ -1,0 +1,13 @@
+import { h } from "preact";
+
+export const ScrollBox = ({ width, height, children }) => (
+  <div
+    style={{
+      width,
+      height,
+      overflow: "scroll"
+    }}
+  >
+    {children}
+  </div>
+);