+++ src/index.js
@@ -1,13 +1,11 @@
-import "babel-polyfill";
-import "./styles/normalize.css";
-import "./styles/style.css";
+import "@babel/polyfill";
+import "!style-loader!css-loader!postcss-loader!sass-loader!./styles/normalize.css";
+import "!style-loader!css-loader!postcss-loader!sass-loader!./styles/style.css";
 import "./components";
 import "./scenery";
 import "./scenes";
 import "./systems/SeaLevel";
-import screenfull from "screenfull";
 
-/* eslint-env node */
 Game.start(false);
 
 /**
@@ -36,11 +34,11 @@ window.addEventListener("resize", scaleGame);
 const button = document.querySelector("button");
 button.addEventListener("click", () => {
   const theater = document.getElementById("theater");
-  screenfull.request(theater);
+  theater.requestFullscreen();
   document.body.classList.add("fullscreen");
   scaleGame();
-  document.addEventListener(screenfull.raw.fullscreenchange, () => {
-    if (!screenfull.isFullscreen) {
+  document.addEventListener("fullscreenchange", () => {
+    if (!document.fullscreenElement) {
       // exit fullscreen code here
       document.body.classList.remove("fullscreen");
       scaleGame();